Refactor: ThermometerTMP36, aggiunta virtual e override

+++ src/devices/button/buttonImpl.cpp
@@ -1,7 +1,7 @@
 #include "Arduino.h"
 #include "buttonImpl.h"
 
-ButtonImpl::ButtonImpl(int pin) : pin(pin) {
+ButtonImpl::ButtonImpl(uint8_t pin): pin(pin) {
     pinMode(pin, INPUT);
 };
