+++ src/devices/led/led.cpp
@@ -1,7 +1,7 @@
-#include "Led.h"
 #include "Arduino.h"
+#include "led.h"
 
-Led::Led(int pin) : pin(pin) {
+Led::Led(uint8_t pin) : pin(pin) {
     // this->pin = pin;
     this->isOn = false;
     pinMode(pin, OUTPUT);
